Products can be deleted

+++ javabackend/src/main/java/javabackend/example/javabackend/repositories/ProductsRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import javabackend.example.javabackend.models.Products;
 
-public interface ProductsRepository extends JpaRepository<Products, Long> {
+public interface ProductsRepository extends JpaRepository<Products, Integer> {
    
     //List<Products> findAll();
